+++ src/01-one/api/aliases.js
@@ -1,0 +1,21 @@
+
+// add method aliases
+const aliases = {
+  'fmt': 'format',
+  'text': 'format',
+  'leapYear': 'isLeapYear',
+  'isLeap': 'isLeapYear',
+  'inDST': 'inDst',
+  'hasDST': 'hasDst',
+  'hours': 'hour',
+  'minutes': 'minute',
+  'seconds': 'second',
+  'minus': 'subtract',
+  'plus': 'add',
+  'isDst': 'inDst',
+  // ['set', '_from'],
+  //   fns.monthName = fns.month
+  // fns.dayName = fns.day
+  // fns.hour12 = fns.hour
+}
+export default aliases