+++ src/01-one/api/change/startOf.js
@@ -74,7 +74,4 @@ export default {
     s = s.subtract(1, 'millisecond')
     return s
   },
-  progress: function (unit) {
-    return {}
-  }
 }