+++ src/01-one/api/progress.js
@@ -1,0 +1,4 @@
+const progress = function (s, unit) {
+  return {}
+}
+export default progress