+++ src/01-one/api/unit/misc.js
@@ -1,4 +1,5 @@
-import tick from './add/tick.js'
+
+import { zeroPad } from './format/_lib.js'
 
 const getter = {
   week: (s) => {
@@ -25,34 +26,6 @@ const setter = {
 export default {
 
 
-  // this one is tricky!
-  day: function (input, fwd) {
-    const { epoch, tz, world } = this
-    const { getDay, getCal } = world.methods
-    let cal = world.methods.getCal(epoch, tz, world)
-    if (input !== undefined) {
-      let day = getDay(cal.year, cal.month, cal.date)
-      if (day === input) {
-        return cal
-      }
-      let diff = input - day
-      // go in a specific direction
-      if (diff < 0 && fwd === true) {
-        diff = 7 + diff
-      } else if (diff > 0 && fwd === false) {
-        diff = diff - 7
-      }
-      return tick(cal, diff, 'date', world)
-    }
-    return world.methods.getDay(cal)
-  },
-  // wednesday/friday
-  dayName: function (input) {
-    const { world } = this
-    if (input !== undefined) { }
-    let n = this.day()
-    return world.model.days[n].longForm
-  },
   monthName: function (input) {
     const { world } = this
     if (input !== undefined) { }
@@ -90,5 +63,56 @@ export default {
       // current: { offset: -5, isDST: false },
       // change: { start: '03/12:02', back: '11/05:02' }
     }
+  },
+
+  time: function (input) {
+    if (input !== undefined) {
+      let { epoch, tz, world } = this
+      let cal = world.methods.getCal(epoch, tz, world)
+      let c = setters.time(input, cal, tz)
+      let e = world.methods.getEpoch(c, tz, this.world)
+      return this._from(e, tz)
+    }
+    return `${this.hour()}:${zeroPad(this.minute())}${this.ampm()}`
+  },
+
+  clone: function () {
+    return this._from(this.epoch, this.tz)
+  },
+
+  isValid: () => true,
+
+  hasDst: function () {
+    let { tz, world } = this
+    return world.zones[tz].dst
+  },
+  isAsleep: function () {
+    return false
+  },
+
+  inDst: function () {
+    let { epoch, tz, world } = this
+    const { getCal, dstChanges } = world.methods
+    // if it doesn't have dst
+    if (!this.hasDst()) {
+      return false
+    }
+    let cal = getCal(epoch, tz, world)
+    let res = dstChanges(tz, cal.year, world)
+    // console.log(res)
+    return true
+  },
+
+  json: function () {
+    let { epoch, tz, world } = this
+    const { getCal, dstChanges } = world.methods
+    let out = getCal(epoch, tz, world)
+    out.epoch = epoch
+    out.tz = tz
+    let z = world.zones[tz] || {}
+    out.hem = z.hem
+    // out.abbrevs = z.shrt
+    out.dst = dstChanges(tz, out.year)
+    return out
   }
 }